// src/app/shared/models/register.model.ts
export class register {
  name?: string;
  email?: string;
  password?: string;
  confirmPassword?: string;
  img?: string;
  content1?: string;
  content2?: string;
  paragraph?: string;
}

export class passwordResponse {
  passwordResponseText?: string;
  passwordResponseImage?: string;
  passwordResponseKey?: string;
}